+++ packages/next/server/dev/next-dev-server.ts
@@ -700,8 +700,16 @@ export default class DevServer extends Server {
     })
   }
 
+  protected getPagesManifest(): undefined {
+    return undefined
+  }
+
   protected getMiddleware(): never[] {
     return []
+  }
+
+  protected getMiddlewareManifest(): undefined {
+    return undefined
   }
 
   protected async hasMiddleware(
